+++ packages/site/CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.0]
+### Added
+- Add JSON-RPC endowment test snap ([#105](https://github.com/MetaMask/test-snaps/pull/105))
+
+## [4.1.2]
+### Changed
+- Show alert on errors ([#99](https://github.com/MetaMask/test-snaps/pull/99))
+
+## [4.1.1]
+### Fixed
+- Remove undefined params in `snap_invokeSnap` ([#100](https://github.com/MetaMask/test-snaps/pull/100))
+
+## [4.1.0]
+### Changed
+- Handle breaking changes with new version of Snaps Platform ([#95](https://github.com/MetaMask/test-snaps/pull/95))
+- Use BIP32 snap for update snap test ([#94](https://github.com/MetaMask/test-snaps/pull/94))
+
 ## [4.0.2]
 ### Changed
 - No changes this release.
@@ -123,7 +140,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Error to index.html
 - Added proper serve command
 
-[Unreleased]: https://github.com/MetaMask/test-snaps/compare/v4.0.2...HEAD
+[Unreleased]: https://github.com/MetaMask/test-snaps/compare/v4.2.0...HEAD
+[4.2.0]: https://github.com/MetaMask/test-snaps/compare/v4.1.2...v4.2.0
+[4.1.2]: https://github.com/MetaMask/test-snaps/compare/v4.1.1...v4.1.2
+[4.1.1]: https://github.com/MetaMask/test-snaps/compare/v4.1.0...v4.1.1
+[4.1.0]: https://github.com/MetaMask/test-snaps/compare/v4.0.2...v4.1.0
 [4.0.2]: https://github.com/MetaMask/test-snaps/compare/v4.0.1...v4.0.2
 [4.0.1]: https://github.com/MetaMask/test-snaps/compare/v4.0.0...v4.0.1
 [4.0.0]: https://github.com/MetaMask/test-snaps/compare/v3.2.0...v4.0.0